# faster xor and != for Bool