@stevengj 2019-03-26
Weirdly, for `UInt8`, `!==` is even faster:
```
julia> h(a) = reduce(!==, a);

julia> @btime h($b); @btime h($i);
  89.135 ns (0 allocations: 0 bytes)
  8.933 ns (0 allocations: 0 bytes)
```
even though `@code_native (!==)(true,false)` and `@code_native (!==)(0x01,0x02)` have identical output.  What gives?