@vtjnash 2019-03-26
Ah, I found some potential codegen optimizations to be able to this automatically–and they made things worse :P

I think we used to have this optimization, but had to remove it because we don't zero-initialize arrays (#30350). This PR possibly just finds a code path that we don't sanitize and codegen correctly.